pyproject.toml
[tool.poetry]
name = "novalabs-backtest"
version = "1.1.10"
description = "Wrappers around Nova Labs utilities focused on safety and testability"
authors = ["Nova Labs <devteam@novalabs.ai>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/Nova-DevTeam/nova-backtest"
repository = "https://github.com/Nova-DevTeam/nova-backtest"
documentation = "https://github.com/Nova-DevTeam/nova-backtest"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
pandas = "^2.0.0"
numpy = "^1.24.2"
pytz = "^2023.3"
python-decouple = "^3.8"
matplotlib ="^3.7.1"
aiohttp = "^3.8.4"
asyncio = "^3.4.3"
psaw = "^0.1.0"
telethon = "^1.27.0"
requests = "^2.28.2"
ta = "^0.10.2"

[tool.poetry.dev-dependencies]
pytest = "^7.2.2"
pre-commit = "^3.0.0"
black = "^23.1.0"
genbadge = "^1.0.0"
pytest-cov = "^4.0.0"
sphinx = "^6.1.3"
sphinx_rtd_theme = "^1.2.0"
setuptools = "^67.6.1"


[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
